listing elements of a link

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -9,7 +9,7 @@
 
 size_t print_listint(const listint_t *h)
 {
-	size_t p;
+	int p;
 
 	if (h == NULL)
 		return (1);
